# storage/app/scribe/openapi.yaml
openapi: 3.0.3
info:
  title: 'AdminSuite API API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
...
paths:
...
  '/api/budgets/fiscal-year/{year}':
    get:
      summary: 'Get budgets by fiscal year'
      operationId: getBudgetsByFiscalYear
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: year
        description: ''
        example: architecto
        required: true
        schema:
          type: string
  '/api/budgets/fund-source/{fundSource}':
    get:
      summary: 'Get budgets by fund source'
      operationId: getBudgetsByFundSource
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: fundSource
        description: ''
        example: architecto
        required: true
        schema:
          type: string
  '/api/budgets/{id}':
    get:
      summary: 'Get budget by ID'
      operationId: getBudgetByID
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    put:
      summary: 'Update budget'
      operationId: updateBudget
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    delete:
      summary: 'Delete budget'
      operationId: deleteBudget
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the budget.'
        example: architecto
        required: true
        schema:
          type: string
  '/api/budgets/{id}/approve':
    put:
      summary: 'Approve budget'
      operationId: approveBudget
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the budget.'
        example: architecto
        required: true
        schema:
          type: string
  '/api/budgets/{id}/activate':
    put:
      summary: 'Activate budget'
      operationId: activateBudget
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the budget.'
        example: architecto
        required: true
        schema:
          type: string
  '/api/budgets/{id}/close':
    put:
      summary: 'Close budget'
      operationId: closeBudget
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the budget.'
        example: architecto
        required: true
        schema:
          type: string
...